유저 닉네임 변경시 댓글, 게시글 닉네임 변경 + 댓글 작성 권한 설정

+++ src/main/java/com/example/newsper/api/CommentApiController.java
@@ -3,7 +3,10 @@ package com.example.newsper.api;
 import com.example.newsper.dto.AddCommentDto;
 import com.example.newsper.dto.CommentDto;
 import com.example.newsper.entity.CommentEntity;
+import com.example.newsper.entity.UserEntity;
+import com.example.newsper.service.ArticleService;
 import com.example.newsper.service.CommentService;
+import com.example.newsper.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -15,7 +18,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Tag(name= "Comment", description = "댓글 API")
 @RestController
@@ -25,13 +30,24 @@ public class CommentApiController {
     @Autowired
     private CommentService commentService;
 
+    @Autowired
+    private UserService userService;
+
+    @Autowired
+    private ArticleService articleService;
+
     @GetMapping("/comment")
     @Operation(summary= "댓글 조회", description= "특정 글의 댓글을 조회합니다.")
     @ApiResponse(responseCode = "200", description = "성공")
-    public ResponseEntity<List<CommentDto>> comments(
+    public ResponseEntity<?> comments(
             @Parameter(description = "게시글ID")
-            @PathVariable Long articleId
+            @PathVariable Long articleId,
+            HttpServletRequest request
     ){
+        String userId = userService.getUserId(request);
+        UserEntity user = userService.findById(userId);
+        if(!commentService.authCheck(articleId, request) || !articleService.isHide(articleService.findById(articleId),user)) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(setErrorCodeBody(-301));
+
         List<CommentDto> dtos = commentService.comments(articleId);
         return ResponseEntity.status(HttpStatus.OK).body(dtos);
     }
@@ -39,12 +55,14 @@ public class CommentApiController {
     @PostMapping("/comment")
     @Operation(summary= "댓글 작성", description= "게시글에 댓글을 작성합니다. 액세스 토큰 필요.")
     @ApiResponse(responseCode = "201", description = "성공")
-    public ResponseEntity<CommentEntity> create(
+    public ResponseEntity<?> create(
             @Parameter(description = "게시글ID")
             @PathVariable Long articleId,
             @RequestBody AddCommentDto dto,
             HttpServletRequest request
     ){
+        if(!commentService.authCheck(articleId,request)) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(setErrorCodeBody(-302));
+
         CommentEntity created = commentService.create(articleId,dto,request);
         return ResponseEntity.status(HttpStatus.CREATED).body(created);
     }
@@ -52,12 +70,16 @@ public class CommentApiController {
     @PatchMapping("/comment/{id}")
     @Operation(summary= "댓글 수정", description= "댓글을 수정합니다.")
     @ApiResponse(responseCode = "200", description = "성공")
-    public ResponseEntity<CommentEntity> update(
+    public ResponseEntity<?> update(
             @Parameter(description = "댓글ID")
             @PathVariable Long id,
             @Parameter(description = "text:String")
-            @RequestBody CommentDto dto
+            @RequestBody CommentDto dto,
+            HttpServletRequest request
     ){
+
+        if(!commentService.writerCheck(commentService.findById(id), request)) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(setErrorCodeBody(-303));
+
         CommentEntity updated = commentService.update(id,dto);
         return ResponseEntity.status(HttpStatus.OK).body(updated);
     }
@@ -65,12 +87,34 @@ public class CommentApiController {
     @DeleteMapping("/comment/{id}")
     @Operation(summary= "댓글 삭제", description= "댓글을 삭제합니다.")
     @ApiResponse(responseCode = "200", description = "성공")
-    public ResponseEntity<CommentEntity> delete(
+    public ResponseEntity<?> delete(
             @Parameter(description = "댓글ID")
             @PathVariable Long id,
-            @PathVariable String articleId
+            HttpServletRequest request
     ){
+
+        if(!commentService.writerCheck(commentService.findById(id), request)) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(setErrorCodeBody(-303));
+        
         commentService.delete(id);
         return ResponseEntity.status(HttpStatus.OK).build();
+    }
+
+    private Map<String, Object> setErrorCodeBody(int code){
+        Map<String, Object> responseBody = new HashMap<>();
+        //responseBody.put("status", HttpStatus.UNAUTHORIZED.value());
+        //responseBody.put("error", HttpStatus.UNAUTHORIZED.getReasonPhrase());
+        responseBody.put("code", code);
+
+        if(code == -101) responseBody.put("message", "로그인 아이디를 찾을 수 없음");
+        else if(code == -102) responseBody.put("message", "로그인 패스워드 불일치");
+        else if(code == -201) responseBody.put("message", "회원 가입 파라미터 누락");
+        else if(code == -202) responseBody.put("message", "회원 가입 이메일 인증 오류");
+        else if(code == -203) responseBody.put("message", "회원 가입 ID 중복");
+        else if(code == -301) responseBody.put("message", "게시판 접근 권한 없음");
+        else if(code == -302) responseBody.put("message", "게시글 쓰기 권한 없음");
+        else if(code == -303) responseBody.put("message", "게시글 수정/삭제 권한 없음");
+        else if(code == -1) responseBody.put("message", "지정되지 않은 에러");
+
+        return responseBody;
     }
 }
